fix: Split into separate class files under Node namespace
