+++ packages/frontend/src/components/Node/PixelPosition.tsx
@@ -1,0 +1,4 @@
+export interface PixelPosition {
+  left: number;
+  top: number;
+}
